// apps/conquest/types/campaign.d.ts
export interface Campaign {
  id: number;
  name: string;
  notes: string;
  playerCount: number;
  playerStartingExperience: number;
  encounter_adjusted_experience: number;
  encounter_experience: number;
  campaigns: array[];
  intro: string;
  mapTerrainType: string;
  player_experience_start: number;
  mapDimensions: string;
  hasWeather: boolean;
  encounters: array[];
  timeOfDay: string;
  objective: string;
  group_alive: boolean;
  player_count: number;
  adventuringDayXpLimit: number;
  xpThresholdsByCharacterLevelEasy: number;
  xpThresholdsByCharacterLevelMedium: number;
  xpThresholdsByCharacterLevelHard: number;
  xpThresholdsByCharacterLevelDeadly: number;
  adventuringDayXpStart: number;
  encounterDifficultyOptions: string;
  adventuringDayXpFinish: number;
  playerExperienceEarnedFromEncounter: number;
  playerExperienceFinish: number;
  levelOfPlayersCharactersFinish: number;
  percentageOfAdventuringDayXpRemaining: number;
}